move the get_version code into own submodules
also splits up the get version logic and introduces a type for the parse function
