+++ src/setuptools_scm/_cli.py
@@ -4,9 +4,9 @@ import argparse
 import os
 import sys
 
-from setuptools_scm import _get_version
 from setuptools_scm import Configuration
 from setuptools_scm._file_finders import find_files
+from setuptools_scm._get_version import _get_version
 from setuptools_scm.discover import walk_potential_roots
